break selectCharacters into modular, manageable files
