+++ src/javascript/render/createAttackBtn.js
@@ -1,0 +1,6 @@
+export let createAttackBtn = () => {
+  let attackBtn = document.createElement("button");
+  attackBtn.setAttribute("id", "attack");
+  attackBtn.innerText = "Attack";
+  return attackBtn;
+};
