+++ src/javascript/render/createBattleGround.js
@@ -1,0 +1,7 @@
+export let createBattleGround = () => {
+  let arena = document.getElementById("arena");
+  let battleGround = document.createElement("div");
+  battleGround.setAttribute("id", "battle-ground");
+  arena.appendChild(battleGround);
+  return battleGround;
+};
